Rename UnknownParentException to UnknownEntityTypeParentException Change constructor arguments from Object to String
